requirements.txt
discord.py-self==2.0.0
python-dotenv==1.0.0
Requests==2.31.0
